+++ lib/atlas/entity/base_entity.rb
@@ -60,6 +60,16 @@ module Atlas
         refresh_validation
       end
 
+      # TO OVERRIDE
+      def can_transform?(field)
+        true
+      end
+
+      # TO OVERRIDE
+      def can_group_by?(field)
+        true
+      end
+
       def to_hash
         keys = dynamic_attributes? ? @parameters.keys : internal_parameters
         values = @parameters.values_at(*keys)