+++ lib/atlas/service/mechanism/filtering.rb
@@ -26,6 +26,7 @@ module Atlas
         private_class_method :normalize_filter
 
         def self.normalize_value(entity, field, value)
+          return value unless entity
           subparameters = entity.instance_subparameters
           return value unless subparameters.keys.include?(field)
           value.send(subparameters[field])
@@ -36,7 +37,7 @@ module Atlas
           conjunction, field, operator = filter_parts[0, 3]
           return false unless CONJUNCTIONS.include?(conjunction)
           return false unless OPERATORS.include?(operator)
-          return true if entity.instance_parameters.include?(field)
+          return true if !entity || entity.instance_parameters.include?(field)
           entity.instance_subparameters.keys.include?(field)
         end
         private_class_method :validate_filter_parts