+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## CHANGELOG
 
+## 0.12.0
+- Changed BaseMongoidRepository to receive model and entity in initialize
+
+## 0.11.0
+- Changed Suporte a múltiplos namespaces de serializers
+
 ## 0.10.4
 - Changed Tamanho do stacktrace nos alertas do Slack
 - Added Identificação do job nas informações adicionais de falha